+++ mflow/__init__.py
@@ -1,4 +1,8 @@
 
 from .mflow import *
 
+#TODO: deprecate the tools name in favor of utils for consistency with all our other projects
+#TODO: can this throw a DeprecationWarning?
+from . import utils as tools
+
