+++ mflow/cli/generate.py
@@ -2,7 +2,7 @@ import argparse
 import json
 
 import mflow
-from mflow.tools import ThroughputStatisticsPrinter
+from mflow.utils import ThroughputStatisticsPrinter
 
 
 def main():
