+++ mflow/mflow.py
@@ -4,7 +4,7 @@ import sys
 import zmq
 
 from mflow.handlers import array_1_0, dheader_1_0, dimage_1_0, dseries_end_1_0, raw_1_0
-from mflow.tools import ConnectionCountMonitor, SocketEventListener, no_clients_timeout_notifier
+from mflow.utils import ConnectionCountMonitor, SocketEventListener, no_clients_timeout_notifier
 
 try:
     import ujson as json
